Data_Labs/DVC_Labs/Lab_1/requirements.txt
dvc
torch
torchvision
scikit-learn
pyyaml
